backend/requirements.txt
fastapi
uvicorn
onnxruntime
opencv-python
numpy
scikit-learn
sqlalchemy
python-multipart
cryptography
matplotlib
tqdm
mediapipe==0.10.14
requests
pillow
PyYAML